fix: align setter with field size

+++ server/src/main/java/org/eclipse/openvsx/publish/PublishingConfig.java
@@ -57,7 +57,7 @@ public class PublishingConfig {
         return maxContentSize;
     }
 
-    public void setMaxContentSize(int maxContentSize) {
+    public void setMaxContentSize(long maxContentSize) {
         this.maxContentSize = maxContentSize;
     }
